refactor : rename form controllers to form fields
