+++ src/pages/join/_components/join-fields/JoinPasswordField.tsx
@@ -1,24 +1,13 @@
-import React from "react";
-
 import { TextField } from "@mui/material";
 import { Controller, useFormContext } from "react-hook-form";
 import { useTranslation } from "react-i18next";
 
-import { LoginFormSchema } from "@/pages/login/_schema/LoginFormSchema";
+import { JoinFormSchema } from "@/pages/join/_schema/JoinFormSchema";
 
-interface PasswordControllerProps {
-  onEnter: () => void;
-}
-
-function LoginPasswordController({ onEnter }: PasswordControllerProps) {
+function JoinPasswordField() {
   const { t } = useTranslation();
-  const { control } = useFormContext<LoginFormSchema>();
 
-  const handleEnterKeyDown = (e: React.KeyboardEvent) => {
-    if (e.key === "Enter") {
-      onEnter();
-    }
-  };
+  const { control } = useFormContext<JoinFormSchema>();
 
   return (
     <Controller
@@ -27,9 +16,8 @@ function LoginPasswordController({ onEnter }: PasswordControllerProps) {
       render={({ field, fieldState: { error } }) => (
         <TextField
           type="password"
-          placeholder={t("login.form.label.PASSWORD")}
+          placeholder={t("join.form.label.PASSWORD")}
           onChange={field.onChange}
-          onKeyDown={handleEnterKeyDown}
           error={!!error}
           sx={{
             width: "100%",
@@ -41,4 +29,4 @@ function LoginPasswordController({ onEnter }: PasswordControllerProps) {
   );
 }
 
-export default LoginPasswordController;
+export default JoinPasswordField;
