+++ src/pages/join/_components/join-fields/JoinUsernameField.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from "react-i18next";
 
 import { JoinFormSchema } from "@/pages/join/_schema/JoinFormSchema";
 
-function JoinUsernameController() {
+function JoinUsernameField() {
   const { t } = useTranslation();
 
   const { control } = useFormContext<JoinFormSchema>();
@@ -36,4 +36,4 @@ function JoinUsernameController() {
   );
 }
 
-export default JoinUsernameController;
+export default JoinUsernameField;
